Removed routes from server.js and created route modules in server/routes folder
